Maybe fixed video on dummy renderer on Xbox Series X|S

+++ Content/SampleVertexShader.hlsl
@@ -1,15 +1,15 @@
 // Per-vertex data used as input to the vertex shader.
 struct VertexShaderInput
 {
-	min16float4 Pos : POSITION;
-	min16float2 Tex : TEXCOORD;
+	float4 Pos : POSITION;
+	float2 Tex : TEXCOORD;
 };
 
 // Per-pixel color data passed through the pixel shader.
 struct PixelShaderInput
 {
-	min16float4 Pos : SV_POSITION;
-	min16float2 Tex : TEXCOORD;
+	float4 Pos : SV_POSITION;
+	float2 Tex : TEXCOORD;
 };
 
 // Simple shader to do vertex processing on the GPU.
